pyproject.toml
[project]
name = "dong-timeline"
version = "0.1.0"
description = "时间咚 - 记录人生/项目的关键节点"
authors = [{name = "gudong", email = "gudong@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "dong-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dong-timeline = "timeline.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
